Move transform script into mapping-transform
